+++ DemoApp/Sources/Views/CallView/CallingView/DetailedCallingView.swift
@@ -151,9 +151,13 @@ struct DetailedCallingView<Factory: ViewFactory>: View {
                 resignFirstResponder()
                 Task {
                     await setPreferredVideoCodec(for: text)
-                    await setEncryptionIfNeeded(for: text)
                     if callAction == .joinCall {
-                        viewModel.joinCall(callType: callType, callId: text)
+                        await setEncryptionIfNeeded(for: text)
+                        viewModel.joinCall(
+                            callType: callType,
+                            callId: text,
+                            encryption: AppEnvironment.EncryptionKeys.shared.encryptionRequest
+                        )
                     } else {
                         if callFlow == .lobby {
                             viewModel.enterLobby(
@@ -162,13 +166,16 @@ struct DetailedCallingView<Factory: ViewFactory>: View {
                                 members: members
                             )
                         } else if callFlow == .joinAndRing {
+                            await setEncryptionIfNeeded(for: text)
                             viewModel.joinAndRingCall(
                                 callType: callType,
                                 callId: text,
                                 members: members,
-                                video: viewModel.callSettings.videoOn
+                                video: viewModel.callSettings.videoOn,
+                                encryption: AppEnvironment.EncryptionKeys.shared.encryptionRequest
                             )
                         } else {
+                            await setEncryptionIfNeeded(for: text)
                             let highScaleHint = AppEnvironment
                                 .highScaleLivestreamPublisherHint
                                 .value
@@ -178,7 +185,8 @@ struct DetailedCallingView<Factory: ViewFactory>: View {
                                 members: members,
                                 ring: callFlow == .ringEvents,
                                 highScaleLivestreamPublisherHint: highScaleHint,
-                                video: viewModel.callSettings.videoOn
+                                video: viewModel.callSettings.videoOn,
+                                encryption: AppEnvironment.EncryptionKeys.shared.encryptionRequest
                             )
                         }
                     }